reorganize normalization funcs
