+++ dynamo/preprocessing/utils.py
@@ -383,169 +383,6 @@ def get_svr_filter(
     return res
 
 
-def sz_util(
-    adata: anndata.AnnData,
-    layer: str,
-    round_exprs: bool,
-    method: Literal["mean-geometric-mean-total", "geometric", "median"],
-    locfunc: Callable,
-    total_layers: List[str] = None,
-    CM: pd.DataFrame = None,
-    scale_to: Union[float, None] = None,
-) -> Tuple[pd.Series, pd.Series]:
-    """Calculate the size factor for a given layer.
-
-    Args:
-        adata: an AnnData object.
-        layer: the layer to operate on.
-        round_exprs: whether the gene expression should be rounded into integers.
-        method: the method used to calculate the expected total reads / UMI used in size factor calculation. Only
-            `mean-geometric-mean-total` / `geometric` and `median` are supported. When `mean-geometric-mean-total` is
-            used, size factors will be calculated using the geometric mean with given mean function. When `median` is
-            used, `locfunc` will be replaced with `np.nanmedian`. When `mean` is used, `locfunc` will be replaced with
-            `np.nanmean`. Defaults to "median".
-        locfunc: the function to normalize the data.
-        total_layers: the layer(s) that can be summed up to get the total mRNA. For example, ["spliced", "unspliced"],
-            ["uu", "ul", "su", "sl"] or ["new", "old"], etc. Defaults to None.
-        CM: the data to operate on, overriding the layer. Defaults to None.
-        scale_to: the final total expression for each cell that will be scaled to. Defaults to None.
-
-    Raises:
-        NotImplementedError: method is invalid.
-
-    Returns:
-        A tuple (sfs, cell_total) where sfs is the size factors and cell_total is the initial cell size.
-    """
-
-    adata = adata.copy()
-
-    if layer == "_total_" and "_total_" not in adata.layers.keys():
-        if total_layers is not None:
-            total_layers, _ = DKM.aggregate_layers_into_total(
-                adata,
-                total_layers=total_layers,
-                extend_layers=False,
-            )
-
-    CM = DKM.select_layer_data(adata, layer) if CM is None else CM
-    if CM is None:
-        return None, None
-
-    if round_exprs:
-        main_debug("rounding expression data of layer: %s during size factor calculation" % (layer))
-        if issparse(CM):
-            CM.data = np.round(CM.data, 0)
-        else:
-            CM = CM.round().astype("int")
-
-    cell_total = CM.sum(axis=1).A1 if issparse(CM) else CM.sum(axis=1)
-    cell_total += cell_total == 0  # avoid infinity value after log (0)
-
-    if method in ["mean-geometric-mean-total", "geometric"]:
-        sfs = cell_total / (np.exp(locfunc(np.log(cell_total))) if scale_to is None else scale_to)
-    elif method == "median":
-        sfs = cell_total / (np.nanmedian(cell_total) if scale_to is None else scale_to)
-    elif method == "mean":
-        sfs = cell_total / (np.nanmean(cell_total) if scale_to is None else scale_to)
-    else:
-        raise NotImplementedError(f"This method {method} is not supported!")
-
-    return sfs, cell_total
-
-
-def get_sz_exprs(
-    adata: anndata.AnnData, layer: str, total_szfactor: Union[str, None] = None
-) -> Tuple[np.ndarray, npt.ArrayLike]:
-    """Get the size factor from an AnnData object.
-
-    Args:
-        adata: an AnnData object.
-        layer: the layer for which to get the size factor.
-        total_szfactor: the key-name for total size factor entry in `adata.obs`. If not None, would override the layer
-            selected. Defaults to None.
-
-    Returns:
-        A tuple (szfactors, CM), where szfactors is the queried size factor and CM is the data of the layer
-        corresponding to the size factor.
-    """
-
-    if layer == "raw":
-        CM = adata.raw.X
-        szfactors = adata.obs[layer + "Size_Factor"].values[:, None]
-    elif layer == "X":
-        CM = adata.X
-        szfactors = adata.obs["Size_Factor"].values[:, None]
-    elif layer == "protein":
-        if "protein" in adata.obsm_keys():
-            CM = adata.obsm[layer]
-            szfactors = adata.obs["protein_Size_Factor"].values[:, None]
-        else:
-            CM, szfactors = None, None
-    else:
-        CM = adata.layers[layer]
-        szfactors = adata.obs[layer + "_Size_Factor"].values[:, None]
-
-    if total_szfactor is not None and total_szfactor in adata.obs.keys():
-        szfactors = adata.obs[total_szfactor][:, None]
-    elif total_szfactor is not None:
-        main_warning("`total_szfactor` is not `None` and it is not in adata object.")
-
-    return szfactors, CM
-
-
-def normalize_mat_monocle(
-    mat: np.ndarray,
-    szfactors: np.ndarray,
-    relative_expr: bool,
-    pseudo_expr: int,
-    norm_method: Callable = np.log1p,
-) -> np.ndarray:
-    """Normalize the given array for monocle recipe.
-
-    Args:
-        mat: the array to operate on.
-        szfactors: the size factors corresponding to the array.
-        relative_expr: whether we need to divide gene expression values first by
-            size factor before normalization.
-        pseudo_expr: a pseudocount added to the gene expression value before
-            log/log2 normalization.
-        norm_method: the method used to normalize data. Defaults to np.log1p.
-
-    Returns:
-        The normalized array.
-    """
-
-    if norm_method == np.log1p:
-        pseudo_expr = 0
-    if relative_expr:
-        mat = mat.multiply(csr_matrix(1 / szfactors)) if issparse(mat) else mat / szfactors
-
-    if pseudo_expr is None:
-        pseudo_expr = 1
-
-    if issparse(mat):
-        mat.data = norm_method(mat.data + pseudo_expr) if norm_method is not None else mat.data
-        if norm_method is not None and norm_method.__name__ == "Freeman_Tukey":
-            mat.data -= 1
-    else:
-        mat = norm_method(mat + pseudo_expr) if norm_method is not None else mat
-
-    return mat
-
-
-def size_factor_normalize(mat: np.ndarray, szfactors: np.ndarray) -> np.ndarray:
-    """perform size factor normalization on the given array.
-
-    Args:
-        mat: the array to operate on.
-        szfactors: the size factors corresponding to the array.
-
-    Returns:
-        The normalized array divided by size factor
-    """
-    return mat.multiply(csr_matrix(1 / szfactors)) if issparse(mat) else mat / szfactors
-
-
 def _Freeman_Tukey(X: np.ndarray, inverse=False) -> np.ndarray:
     """perform Freeman-Tukey transform or inverse transform on the given array.
